requirements.txt
flask
pandas
python-dotenv
psycopg2-binary

# pip install -r requirements.txt